Find existing correspond

Unlike get_correspond() functions, this function will find existing correspond if the specified year has no correspond codes. Correspond codes can be empty if nothing has changed in that specific year and need to get from previous year or even year before before previous year etc..etc.. This function is needed when running cast_geo().

find_correspond(type, correspond, from)

Arguments

type
Higher granularity from specified correspond arg.

correspond
Lower granularity from the specified type arg.

from
Specify the starting year for range period. Current year is the default.

Value

A dataset of class data.table representing the lower geographical level codes and their corresponding higher geographical levels. For example for codes on enumeration areas and their corresponding codes for municipalities or town.