Get geo code that are split after code change

Get geo code that are split after code change

Usage

track_split(
  type = c("fylke", "okonomisk", "kommune", "bydel", "levekaar", "grunnkrets"),
  from = NULL,
  to = NULL,
  names = TRUE
)

Arguments

type

Type of regional granularity ie. fylke, kommune etc.

from

Specify the starting year for range period. Current year is the default.

to

Specify the year to end the range period. Current year is used when not specified.

names

Include names. Default is TRUE

Value

Dataset of class data.table with column split showing the number of time the codes have been split to

Examples

 dt <- track_split("kommune", 2018, 2020)
#> Please check splitting geo codes with `track_split()` for possible error
#> Use `fix = TRUE` to implement the way we do it in our work. Please read the documentation